+++ test/Buzz/Test/Integration/BaseIntegrationTest.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Buzz\Test\Integration;
+
+use Http\Client\Tests\HttpClientTest;
+
+abstract class BaseIntegrationTest extends HttpClientTest
+{
+    public function testSendWithInvalidUri()
+    {
+        $this->markTestSkipped('We do not support HTTPlugs exceptions');
+    }
+}